Working front-end form for move submission
On the new_ivanhoe_move.php file, I've created a working form that will accept title and content, and uses wp_insert_post to make that into a move that shows up in the dashboard and all. However, the parent id does not yet seem to be working, so we need to continue working on that.