wp_redirect attempt
We pulled in the permalink of the game page and passed it into the form to submit a new move. We then tried to use a wp_redirect in the top if statement to send the user back to the game page with which the move is associated. Unfortunately, this did not work, and after submitting the move it returned a blank page. I have commented out the code for the moment.