Removes insert-media class.
This seems to mess with the Add Media button, causing the modal window for
adding media to activate regardless of where you click on the form. Besides, the
content spans across all columns anyways, so there isn't a need for a
specific rule to span all the columns.

If there is a need to add a style hook in the form of a class, we should come up
with a value other than 'insert-media', since WP seems to use that value for
activating the media modal window.