Adjusts widths and padding in lots of places.
Makes teh whole page a bit wider to account for two columns of potentially long
text. Reduces the number of places with padding to make the spacing more
consistent with the Susy grid we have set up.

Also centers our logo in the footer because, I dunno, it feels better.