Markup and style changes for author page.
Adds a bit more structure to the author page. Specifically, this wraps each role
in an article tag, and using a top-level heading for the role name itself. It
also includes the role thumbnail.

Lastly, this implements a two-column layout for the profile page, with the left
column featuring a list of user roles, and the right column showing the user's
description if one exists.