Adds other WordPress i18n function to xgettext.
Passes additional WordPress i18n functions to the xgettext task in the build
script, so it knows to look for strings in those functions. The build script now
looks for:

* __()
* _e()
* _x()

There are additional functions for i18n available in WordPress, but those are
not currently used in our theme. We might eventually want add all of them for
coverage.

Refs #134.