Wraps page titles with i18n function.
This fixes an issue I encountered where, having a translation of the Ivanhoe
theme loaded, the checks were not working, because the titles for these pages
were saved in a language other than English.

Even with this update, we should note that these checks are pretty vulnerable,
since users can easily change the titles of these pages by just editing the
page in their WP admin. Perhaps we should figure out some other way, maybe
with a custom field on the page, to indicate which page is being used for these
forms.