Fixed rationales query.
The main problem was that you can't have duplicate keys in an array.

Also, for some reason I couldn't get it to query properly on the role
ID, so instead I queried on the author (since the role author and the
rationale author will always be the same and since there's only one role
per user).